feat: move characters.json
