+++ db/seeds/001-characters.js
@@ -1,3 +1,5 @@
+const hpData = require('../../data/characters.json')
+
 /**
  * @param { import("knex").Knex } knex
  * @returns { Promise<void> }
@@ -5,196 +7,5 @@
 exports.seed = async function (knex) {
     // Deletes ALL existing entries
     await knex('characters').del()
-    await knex('characters').insert([
-        {
-            id: 1,
-            name: 'Harry Potter',
-            species: 'human',
-            gender: 'male',
-            house: 'Gryffindor',
-            dateOfBirth: '31-07-1980',
-            yearOfBirth: 1980,
-            wizard: true,
-            ancestry: 'half-blood',
-            eyeColour: 'green',
-            hairColour: 'black',
-            patronus: 'stag',
-            hogwartsStudent: true,
-            hogwartsStaff: false,
-            actor: 'Daniel Radcliffe',
-            alive: true,
-            image: 'https://ik.imagekit.io/hpapi/harry.jpg',
-        },
-        {
-            id: 2,
-            name: 'Hermione Granger',
-            species: 'human',
-            gender: 'female',
-            house: 'Gryffindor',
-            dateOfBirth: '19-09-1979',
-            yearOfBirth: 1979,
-            wizard: true,
-            ancestry: 'muggleborn',
-            eyeColour: 'brown',
-            hairColour: 'brown',
-            patronus: 'otter',
-            hogwartsStudent: true,
-            hogwartsStaff: false,
-            actor: 'Emma Watson',
-            alive: true,
-            image: 'https://ik.imagekit.io/hpapi/hermione.jpeg',
-        },
-        {
-            id: 3,
-            name: 'Ron Weasley',
-            species: 'human',
-            gender: 'male',
-            house: 'Gryffindor',
-            dateOfBirth: '01-03-1980',
-            yearOfBirth: 1980,
-            wizard: true,
-            ancestry: 'pure-blood',
-            eyeColour: 'blue',
-            hairColour: 'red',
-            patronus: 'Jack Russell terrier',
-            hogwartsStudent: true,
-            hogwartsStaff: false,
-            actor: 'Rupert Grint',
-            alive: true,
-            image: 'https://ik.imagekit.io/hpapi/ron.jpg',
-        },
-        {
-            id: 4,
-            name: 'Draco Malfoy',
-            species: 'human',
-            gender: 'male',
-            house: 'Slytherin',
-            dateOfBirth: '05-06-1980',
-            yearOfBirth: 1980,
-            wizard: true,
-            ancestry: 'pure-blood',
-            eyeColour: 'grey',
-            hairColour: 'blonde',
-            patronus: '',
-            hogwartsStudent: true,
-            hogwartsStaff: false,
-            actor: 'Tom Felton',
-            alive: true,
-            image: 'https://ik.imagekit.io/hpapi/draco.jpg',
-        },
-        {
-            id: 5,
-            name: 'Minerva McGonagall',
-            species: 'human',
-            gender: 'female',
-            house: 'Gryffindor',
-            dateOfBirth: '04-10-1925',
-            yearOfBirth: 1925,
-            wizard: true,
-            ancestry: 'half-blood',
-            eyeColour: '',
-            hairColour: 'black',
-            patronus: 'tabby cat',
-            hogwartsStudent: false,
-            hogwartsStaff: true,
-            actor: 'Dame Maggie Smith',
-            alive: true,
-            image: 'https://ik.imagekit.io/hpapi/mcgonagall.jpg',
-        },
-        {
-            id: 6,
-            name: 'Cedric Diggory',
-            species: 'human',
-            gender: 'male',
-            house: 'Hufflepuff',
-            dateOfBirth: null,
-            yearOfBirth: 1977,
-            wizard: true,
-            ancestry: '',
-            eyeColour: 'grey',
-            hairColour: 'brown',
-            patronus: '',
-            hogwartsStudent: true,
-            hogwartsStaff: false,
-            actor: 'Robert Pattinson',
-            alive: false,
-            image: 'https://ik.imagekit.io/hpapi/cedric.png',
-        },
-        {
-            id: 7,
-            name: 'Cho Chang',
-            species: 'human',
-            gender: 'female',
-            house: 'Ravenclaw',
-            dateOfBirth: '07-04-1979',
-            yearOfBirth: 1979,
-            wizard: true,
-            ancestry: '',
-            eyeColour: 'brown',
-            hairColour: 'black',
-            patronus: 'swan',
-            hogwartsStudent: true,
-            hogwartsStaff: false,
-            actor: 'Katie Leung',
-            alive: true,
-            image: 'https://ik.imagekit.io/hpapi/cho.jpg',
-        },
-        {
-            id: 8,
-            name: 'Severus Snape',
-            species: 'human',
-            gender: 'male',
-            house: 'Slytherin',
-            dateOfBirth: '09-01-1960',
-            yearOfBirth: 1960,
-            wizard: true,
-            ancestry: 'half-blood',
-            eyeColour: 'black',
-            hairColour: 'black',
-            patronus: 'doe',
-            hogwartsStudent: false,
-            hogwartsStaff: true,
-            actor: 'Alan Rickman',
-            alive: false,
-            image: 'https://ik.imagekit.io/hpapi/snape.jpg',
-        },
-        {
-            id: 9,
-            name: 'Rubeus Hagrid',
-            species: 'half-giant',
-            gender: 'male',
-            house: 'Gryffindor',
-            dateOfBirth: '06-12-1928',
-            yearOfBirth: 1928,
-            wizard: true,
-            ancestry: 'half-blood',
-            eyeColour: 'black',
-            hairColour: 'black',
-            patronus: '',
-            hogwartsStudent: false,
-            hogwartsStaff: true,
-            actor: 'Robbie Coltrane',
-            alive: true,
-            image: 'https://ik.imagekit.io/hpapi/hagrid.png',
-        },
-        {
-            id: 10,
-            name: 'Neville Longbottom',
-            species: 'human',
-            gender: 'male',
-            house: 'Gryffindor',
-            dateOfBirth: '30-07-1980',
-            yearOfBirth: 1980,
-            wizard: true,
-            ancestry: 'pure-blood',
-            eyeColour: '',
-            hairColour: 'blonde',
-            patronus: 'Non-Corporeal',
-            hogwartsStudent: true,
-            hogwartsStaff: false,
-            actor: 'Matthew Lewis',
-            alive: true,
-            image: 'https://ik.imagekit.io/hpapi/neville.jpg',
-        },
-    ])
+    await knex('characters').insert(hpData.characters)
 }
